+++ src/OrdersApi/Contracts/OrderLineDto.cs
@@ -4,4 +4,4 @@ namespace OrdersApi.Contracts;
 /// <param name="ProductId">Идентификатор товара</param>
 /// <param name="Quantity">Количество (минимум 1)</param>
 /// <param name="UnitPrice">Цена за единицу</param>
-public record OrderLineDto(Guid ProductId, int Quantity, decimal UnitPrice);
+public record OrderLineDto(Guid ProductId, int Quantity, double UnitPrice);
