+++ src/OrdersApi/Contracts/OrderSummaryDto.cs
@@ -12,5 +12,5 @@ public record OrderSummaryDto(
     Guid Id,
     Guid CustomerId,
     OrderStatus Status,
-    decimal Total,
+    double Total,
     DateTimeOffset CreatedAt);
