+++ src/ProductsApi/Contracts/CreateProductRequest.cs
@@ -6,7 +6,7 @@ namespace ProductsApi.Contracts;
 public record CreateProductRequest(
     [Required, MinLength(1)] string Name,
     string? Description,
-    [Range(0, double.MaxValue)] decimal Price,
+    [Range(0, double.MaxValue)] double Price,
     [Range(0, int.MaxValue)] int StockQuantity,
     Guid CategoryId,
     string? Sku);
